requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
reportlab>=4.0.0
Pillow>=10.0.0
email-validator>=2.0.0
pymongo==4.4.0
werkzeug>=2.3.7
appwrite==4.0.0
